i regretted. having an extra layer is useful in cases,
for example, we could override the method from struct,
otherwise we can't. also, constant resolution could be
another issue. so i think it's better to keep it consistent
rather than doing this and that in different cases for
only a very small portion of advantage.

Revert "reduce a number of ancestor layer"

This reverts commit f3d4cc461ae0d62e05b544eef92526e0ea61d3bf.